Move terminalRemote and parts of native contribution to sandbox
Part of #117257
