+++ src/vs/workbench/contrib/terminal/browser/terminalInstanceService.ts
@@ -10,7 +10,7 @@ import type { SearchAddon as XTermSearchAddon } from 'xterm-addon-search';
 import type { Unicode11Addon as XTermUnicode11Addon } from 'xterm-addon-unicode11';
 import type { WebglAddon as XTermWebglAddon } from 'xterm-addon-webgl';
 import { IProcessEnvironment } from 'vs/base/common/platform';
-import { Emitter, Event } from 'vs/base/common/event';
+import { Emitter } from 'vs/base/common/event';
 import { registerSingleton } from 'vs/platform/instantiation/common/extensions';
 import { Disposable } from 'vs/base/common/lifecycle';
 
@@ -22,11 +22,7 @@ let WebglAddon: typeof XTermWebglAddon;
 export class TerminalInstanceService extends Disposable implements ITerminalInstanceService {
 	public _serviceBrand: undefined;
 
-	readonly onPtyHostExit = Event.None;
-	readonly onPtyHostUnresponsive = Event.None;
-	readonly onPtyHostResponsive = Event.None;
-	readonly onPtyHostRestart = Event.None;
-	private readonly _onRequestDefaultShellAndArgs = this._register(new Emitter<IDefaultShellAndArgsRequest>());
+	protected readonly _onRequestDefaultShellAndArgs = this._register(new Emitter<IDefaultShellAndArgsRequest>());
 	readonly onRequestDefaultShellAndArgs = this._onRequestDefaultShellAndArgs.event;
 
 	public async getXtermConstructor(): Promise<typeof XTermTerminal> {
